SafetyNet Attestation Support Logic
This implements SafetyNet Attestation support without UI part. SafetyNet
Attestation is used to protect a user as a part of abuse detection
system to help determine whether servers are interacting with genuine app
running on a genuine Android device.

"The SafetyNet Attestation API provides a cryptographically-signed
attestation, assessing the device's integrity. In order to create the
attestation, the API examines the device's software and hardware
environment, looking for integrity issues, and comparing it with the
reference data for approved Android devices. The generated attestation
is bound to the nonce that the caller app provides. The attestation also
contains a generation timestamp and metadata about the requesting app."

There is two missing parts that have to be implemented before release:
1. There is missing UI to notify a user about SafetyNet result
2. In GetSafetyNetNonceDataUseCase, OpenTraceRepository
retrieveTemporaryID methods returns a null. In the result limiting
generated nonce quality